+++ data/fileio/CoreAudioFileReader.cpp
@@ -59,7 +59,7 @@ codestr(OSStatus err)
 CoreAudioFileReader::CoreAudioFileReader(FileSource source,
                                          DecodeMode decodeMode,
                                          CacheMode mode,
-                                         size_t targetRate,
+                                         int targetRate,
                                          ProgressReporter *reporter) :
     CodedAudioFileReader(mode, targetRate),
     m_source(source),
